+++ samplesheets/src/components/Overview.vue
@@ -5,23 +5,18 @@
         <h4>Investigation Details</h4>
       </div>
       <div class="card-body">
-        <dl class="row pb-0">
-          <dt class="col-md-3">Identifier</dt>
-          <dd class="col-md-9">{{ app.sodarContext['investigation']['identifier'] }}</dd>
-        </dl>
-        <dl class="row pb-0">
-          <dt class="col-md-3">Title</dt>
-          <dd class="col-md-9">{{ app.sodarContext['investigation']['title'] }}</dd>
-        </dl>
-        <dl v-if="app.sodarContext['investigation']['description']"
-            class="row pb-0">
-          <dt class="col-md-3">Description</dt>
-          <dd class="col-md-9">{{ app.sodarContext['investigation']['description'] }}</dd>
-        </dl>
-        <dl class="row pb-0">
-          <dt class="col-md-3">Parser Version</dt>
-          <dd class="col-md-9">{{ app.sodarContext['parser_version']}}</dd>
-        </dl>
+        <list-row :legend="'Identifier'"
+                  :value="app.sodarContext['investigation']['identifier']">
+        </list-row>
+        <list-row :legend="'Title'"
+                  :value="app.sodarContext['investigation']['title']">
+        </list-row>
+        <list-row :legend="'Description'"
+                  :value="app.sodarContext['investigation']['description']">
+        </list-row>
+        <list-row :legend="'Parser Version'"
+                  :value="app.sodarContext['parser_version']">
+        </list-row>
         <dl class="row pb-0">
           <dt class="col-md-3">Configuration</dt>
           <dd class="col-md-9">
@@ -50,12 +45,9 @@
           </dd>
         </dl>
         <span v-for="(val, key, index) in
-                   app.sodarContext['investigation']['comments']"
-            :key="index">
-          <dl class="row pb-0">
-            <dt class="col-md-3">{{ key }}</dt>
-            <dd class="col-md-9">{{ val }}</dd>
-          </dl>
+                     app.sodarContext['investigation']['comments']"
+              :key="index">
+          <list-row :legend="key" :value="val"></list-row>
         </span>
       </div>
     </div>
@@ -68,22 +60,15 @@
         <h4>Study Details: {{ studyInfo['display_name'] }}</h4>
       </div>
       <div class="card-body">
-        <dl class="row pb-0">
-          <dt class="col-md-3">Title</dt>
-          <dd class="col-md-9">{{ studyInfo['display_name']}}</dd>
-        </dl>
-        <dl v-if="studyInfo['description']"
-            class="row pb-0">
-          <dt class="col-md-3">Description</dt>
-          <dd class="col-md-9">{{ studyInfo['description']}}</dd>
-        </dl>
+        <list-row :legend="'Title'"
+                  :value="studyInfo['display_name']">
+        </list-row>
+        <list-row :legend="'Description'"
+                  :value="studyInfo['description']">
+        </list-row>
         <span v-for="(val, key, index) in studyInfo['comments']"
               :key="index">
-          <dl v-if="val"
-              class="row pb-0">
-            <dt class="col-md-3">{{ key }}</dt>
-            <dd class="col-md-9">{{ val }}</dd>
-          </dl>
+          <list-row :legend="key" :value="val"></list-row>
         </span>
       </div>
     </div>
@@ -127,8 +112,13 @@
 </template>
 
 <script>
+import ListRow from './ListRow.vue'
+
 export default {
   name: 'Overview',
+  components: {
+    ListRow
+  },
   props: [
     'app'
   ],